+++ tests/test_clusterings/spectral_clustering/spectral_clustering.py
@@ -5,7 +5,7 @@ from pymilo.utils.data_exporter import prepare_simple_clustering_datasets
 
 MODEL_NAME = "Spectral Clustering"
 
-def spectral_clustering():    
+def spectral_clustering():
     x, y = prepare_simple_clustering_datasets()
     spectral_clustering = SpectralClustering(random_state=5).fit(x, y)
     pymilo_clustering_test(spectral_clustering, MODEL_NAME, x)